+++ src/ros2/ros2.hpp
@@ -29,7 +29,7 @@ class Node
 
     void recreate(const char* node_name = "ros2_xrcedds_participant",unsigned int client_key=0xAABBCCDD);
     void createWallTimer(uint32_t msec, CallbackFunc callback, void* callback_arg, PublisherHandle* pub);
-    void createWallFreq(uint32_t hz, CallbackFunc callback, void* callback_arg, PublisherHandle* pub);
+    void createWallFreq(float hz, CallbackFunc callback, void* callback_arg, PublisherHandle* pub);
     void runPubCallback();
     void runSubCallback(uint16_t reader_id, void* serialized_msg);
